requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-Login==0.6.3
python-dotenv==1.0.1